+++ SmartPower/CHANGELOG.md
@@ -1,4 +1,5 @@
 # v1.9.0 (started on 8/8/2024)
+* Move power generator UI fragment to the middle of the panel.
 
 # v1.8.0 (8 Aug 2024)
 * Migrate to Update 6 mods model. Incompatible with the previous versions of Timberborn!